Move environment badge and utils to modules sync
