+++ src/modules/sync/components/environment-badge.tsx
@@ -1,7 +1,7 @@
 import { __ } from '@wordpress/i18n';
 import { Badge } from 'src/components/badge';
 import { cx } from 'src/lib/cx';
-import { getEnvironmentLabel } from 'src/lib/environment-utils';
+import { getEnvironmentLabel } from 'src/modules/sync/lib/environment-utils';
 import type { SyncSite } from 'src/hooks/use-fetch-wpcom-sites/types';
 
 export type EnvironmentType = string;
